Alterado tipo do atributo de Worker

+++ api-worker/src/main/java/com/arieltintel/apiworker/entities/Worker.java
@@ -10,6 +10,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -24,6 +25,6 @@ public class Worker {
     private Long id;
 
     private String name;
-    private Double dailyIncome;
+    private BigDecimal dailyIncome;
 
 }
